refactor out mergecompactor.go
