+++ pkg/bloomcompactor/bloomcompactor.go
@@ -27,8 +27,6 @@ package bloomcompactor
 import (
 	"context"
 	"fmt"
-	"github.com/grafana/loki/pkg/logproto"
-	"github.com/grafana/loki/pkg/storage/chunk"
 	"math"
 	"os"
 	"time"
@@ -540,115 +538,13 @@ func (c *Compactor) runCompact(ctx context.Context, logger log.Logger, job Job, 
 				}
 			}
 
-			var populate = func(series *v1.Series, bloom *v1.Bloom) error {
-				bloomForChks := v1.SeriesWithBloom{
-					Series: series,
-					Bloom:  bloom,
-				}
+			// FIXME with localdst
+			blockPath := filepath.Join("foobar", "merge-out")
 
-				// Satisfy types for chunks
-				chunkRefs := make([]chunk.Chunk, len(series.Chunks))
-				for i, chk := range series.Chunks {
-					chunkRefs[i] = chunk.Chunk{
-						ChunkRef: logproto.ChunkRef{
-							Fingerprint: uint64(series.Fingerprint),
-							UserID:      meta.TenantID,
-							From:        chk.Start,
-							Through:     chk.End,
-							Checksum:    chk.Checksum,
-						},
-					}
-				}
-
-				chks, err := storeClient.chunk.GetChunks(ctx, chunkRefs)
-				if err != nil {
-					level.Error(logger).Log("msg", "error downloading chunks", "err", err)
-					return err
-				}
-				_ = bt.PopulateSeriesWithBloom(&bloomForChks, chks)
-
-				return nil
-			}
-
-			// Satisfy types for series
-			seriesFromSeriesMeta := make([]*v1.Series, len(job.seriesMetas))
-
-			for i, s := range job.seriesMetas {
-				crefs := make([]v1.ChunkRef, len(s.chunkRefs))
-				for j, chk := range s.chunkRefs {
-					crefs[j] = v1.ChunkRef{
-						Start:    model.Time(chk.MinTime),
-						End:      model.Time(chk.MaxTime),
-						Checksum: chk.Checksum,
-					}
-				}
-				seriesFromSeriesMeta[i] = &v1.Series{
-					Fingerprint: s.seriesFP,
-					Chunks:      crefs,
-				}
-			}
-			seriesIter := v1.NewSliceIter(seriesFromSeriesMeta)
-
-			// TODO: Make blockIters an actual, lazy iterator
-			// Download existing blocks that needs compaction
-			blockIters := make([]v1.PeekingIterator[*v1.SeriesWithBloom], len(blocksToUpdate))
-			for i, b := range blocksToUpdate {
-				lazyBlock, err := c.bloomShipperClient.GetBlock(ctx, b)
-				if err != nil {
-					level.Error(logger).Log("msg", "error downloading block", "err", err)
-					return err
-				}
-
-				// TODO defer removing this blockpath
-				blockPath, err := ExtractBlock(&lazyBlock, c.cfg.WorkingDirectory)
-				if err != nil {
-					level.Error(logger).Log("msg", "error extracting block", "err", err)
-					return err
-				}
-
-				reader := v1.NewDirectoryBlockReader(blockPath)
-				block := v1.NewBlock(reader)
-				blockQuerier := v1.NewBlockQuerier(block)
-
-				blockIters[i] = v1.NewPeekingIter[*v1.SeriesWithBloom](blockQuerier)
-			}
-
-			mergeBuilder := v1.NewMergeBuilder(
-				blockIters,
-				seriesIter,
-				populate)
-			blockPath := filepath.Join(localDst, "merge-out")
-
-			mergeBlockBuilder, err := NewPersistentBlockBuilder(blockPath, blockOptions)
+			mergedBlock, err := mergeCompactChunks(ctx, logger, c.bloomShipperClient, storeClient, bt, job, blockOptions, blocksToUpdate, c.cfg.WorkingDirectory, blockPath)
 			if err != nil {
-				level.Error(logger).Log("msg", "creating block builder", "err", err)
+				level.Error(logger).Log("msg", "failed to merge existing blocks with new chunks", "err", err)
 				return err
-			}
-			checksum, err := mergeBlockBuilder.mergeBuild(mergeBuilder)
-			if err != nil {
-				level.Error(logger).Log("msg", "failed merging the blooms", "err", err)
-				return err
-			}
-			data, err := mergeBlockBuilder.Data()
-			if err != nil {
-				level.Error(logger).Log("msg", "failed reading bloom data", "err", err)
-				return err
-			}
-
-			mergedBlock := bloomshipper.Block{
-				BlockRef: bloomshipper.BlockRef{
-					Ref: bloomshipper.Ref{
-						TenantID:       job.tenantID,
-						TableName:      job.tableName,
-						MinFingerprint: uint64(job.minFp),
-						MaxFingerprint: uint64(job.maxFp),
-						StartTimestamp: int64(job.from),
-						EndTimestamp:   int64(job.through),
-						Checksum:       checksum,
-					},
-					IndexPath: job.indexPath,
-				},
-				Data: data,
 			}
 
 			archivePath := filepath.Join(c.cfg.WorkingDirectory, uuid.New().String())
